Flex, PHP, and Mobile
Big update on this one. Here's what changed:

-Rewrote the layout slightly to use flex instead of bootstrap. This
meant I could completely remove bootstrap altogether.
-Grid layout was adjusted so now the photos will stretch and shrink to
fit the screen up to their max width.
-Files now use PHP includes to load instead of using jQuery to load.
This should be faster and makes much more sense
-Due to the PHP change, I had to rewrite how the photogallery and the
grid go back and forth. It works SO much better now. The gallery
framework is loaded on pageload, but with no photos in it. When a grid
item is selected, it loads the image, changes the visibility and adjusts
the opacity for a smooth transition. the opposite is then done when exit
is pressed.
-Finally, I have added in mobile controls. Now when you swipe on the
gallery page, the photos will transition as expected. Everyone I showed
this to was trying to do that anyway so it makes sense to implement.

Needs:

-I need to figure out a way to implement exiting from the gallery when
on mobile. The escape button works for desktop but not mobile.